refactor : ProjectBoard 서비스 분리 (manage / view)
